* sn2princ.c (krb5_sname_to_principal): Cast tolower argument to int
	* hst_realm.c (krb5_get_host_realm): Cast argument to toupper and
	tolower to int.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13576 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/os/hst_realm.c
@@ -265,7 +265,7 @@ krb5_get_host_realm(context, host, realmsp)
 
     for (cp = local_host; *cp; cp++) {
 	if (isupper((int) (*cp)))
-	    *cp = tolower(*cp);
+	    *cp = tolower((int) *cp);
     }
     l = strlen(local_host);
     /* strip off trailing dot */
@@ -348,7 +348,7 @@ krb5_get_host_realm(context, host, realmsp)
             /* Assume the realm name is upper case */
             for (cp = realm; *cp; cp++)
                 if (islower((int) (*cp)))
-                    *cp = toupper(*cp);
+                    *cp = toupper((int) *cp);
         } else {    
             /* We are defaulting to the local realm */
             retval = krb5_get_default_realm(context, &realm);
